correccion en descuento creacion de orden

+++ ecommerce/src/main/java/com/uade/ecommerce/controller/OrdenController.java
@@ -1,5 +1,6 @@
 package com.uade.ecommerce.controller;
 
+import java.math.BigDecimal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -38,7 +39,7 @@ public class OrdenController {
         public Integer usuarioId; // Se añadió un ID de usuario al request
         public Integer direccionId; // null para retiro en tienda
         public List<ItemCompraRequest> items;
-        public Integer descuento;
+        public BigDecimal descuento;
     }
 
     @Data
